+++ naz/log.py
@@ -204,9 +204,10 @@ class BreachHandler(handlers.MemoryHandler):
     def __init__(
         self,
         flushLevel: int = logging.WARNING,
-        capacity: int = 10_000,
+        capacity: int = 1_000,
         target: logging.Handler = logging.StreamHandler(),
         flushOnClose: bool = False,
+        heartbeatInterval: typing.Union[None, float] = None,
     ) -> None:
         """
         Parameters:
@@ -214,7 +215,17 @@ class BreachHandler(handlers.MemoryHandler):
             capacity: the maximum number of log records to store in the ring buffer
             target: the ultimate `log handler <https://docs.python.org/3.6/library/logging.html#logging.Handler>`_ that will be used.
             flushOnClose: whether to flush the buffer when the handler is closed even if the flush level hasn't been exceeded
+            heartbeatInterval: can be a float or None. If it is a float, then a heartbeat log record will be emitted every :py:attr:`~heartbeatInterval` seconds.
+                               If it is None(the default), then no heartbeat log record is emitted.
+                               If you do decide to set it, a good value is at least 1800(ie 30 minutes).
         """
+        self._validate_args(
+            flushLevel=flushLevel,
+            capacity=capacity,
+            target=target,
+            flushOnClose=flushOnClose,
+            heartbeatInterval=heartbeatInterval,
+        )
         # call `logging.handlers.MemoryHandler` init
         super(BreachHandler, self).__init__(  # type: ignore
             capacity=capacity,
@@ -228,9 +239,78 @@ class BreachHandler(handlers.MemoryHandler):
         # assuming each log record is 250 bytes, then the maximum
         # memory used by `buffer` will always be == 250*10_000/(1000*1000) == 2.5MB
 
+        self.heartbeatInterval = heartbeatInterval
+        if self.heartbeatInterval:
+            self.heartbeatInterval = heartbeatInterval  # seconds
+            self._s_time = time.monotonic()
+
+        self.target.setLevel(logging.DEBUG)  # type: ignore
+
     def shouldFlush(self, record: logging.LogRecord) -> bool:
         """
         Check for record at the flushLevel or higher.
         Implementation is mostly taken from `logging.handlers.MemoryHandler`
         """
+        self._heartbeat()
         return record.levelno >= self.flushLevel  # type: ignore # pytype: disable=attribute-error
+
+    def _heartbeat(self):
+        if not self.heartbeatInterval:
+            return
+
+        # check if `heartbeatInterval` seconds have passed.
+        # if they have, emit a heartbeat log record to the target handler
+        _now = time.monotonic()
+        _diff = _now - self._s_time
+        if _diff >= self.heartbeatInterval:
+            self._s_time = _now
+            # see: https://docs.python.org/3/library/logging.html#logging.LogRecord
+            record = logging.makeLogRecord(
+                {
+                    "level": logging.INFO,
+                    "name": "BreachHandler",
+                    "pathname": ".../naz/naz/log.py",
+                    "func": "BreachHandler._heartbeat",
+                    "msg": {
+                        "event": "naz.BreachHandler.heartbeat",
+                        "heartbeatInterval": self.heartbeatInterval,
+                    },
+                }
+            )
+            self.target.emit(record=record)  # pytype: disable=attribute-error
+
+    def _validate_args(
+        self,
+        flushLevel: int,
+        capacity: int,
+        target: logging.Handler,
+        flushOnClose: bool,
+        heartbeatInterval: typing.Union[None, float],
+    ):
+        if not isinstance(flushLevel, int):
+            raise ValueError(
+                "`flushLevel` should be of type:: `int` You entered: {0}".format(type(flushLevel))
+            )
+        if not isinstance(capacity, int):
+            raise ValueError(
+                "`capacity` should be of type:: `int` You entered: {0}".format(type(capacity))
+            )
+
+        if not isinstance(target, logging.Handler):
+            raise ValueError(
+                "`target` should be of type:: `logging.Handler` You entered: {0}".format(
+                    type(target)
+                )
+            )
+        if not isinstance(flushOnClose, bool):
+            raise ValueError(
+                "`flushOnClose` should be of type:: `bool` You entered: {0}".format(
+                    type(flushOnClose)
+                )
+            )
+        if not isinstance(heartbeatInterval, (type(None), float)):
+            raise ValueError(
+                "`heartbeatInterval` should be of type:: `None` or `float` You entered: {0}".format(
+                    type(heartbeatInterval)
+                )
+            )